addressed feedback. Moved files.
