+++ stablehlo/CMakeLists.txt
@@ -12,9 +12,9 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-add_subdirectory(compatibility)
 add_subdirectory(dialect)
 add_subdirectory(integrations)
 add_subdirectory(reference)
 add_subdirectory(tests)
 add_subdirectory(tools)
+add_subdirectory(transforms)
