+++ stablehlo/dialect/Base.td
@@ -43,7 +43,7 @@ def HLO_Complex : Complex<AnyTypeOf<[F32, F64]>>;
 //===----------------------------------------------------------------------===//
 
 // TODO(b/230381284): Upstream width-specific uniform quantized element types.
-class UniformQuantizedSignedInt<int width>
+class StableHLO_UniformQuantizedSignedInt<int width>
   : Type<Or<[
       And<[CPred<"$_self.isa<mlir::quant::UniformQuantizedType>()">,
            CPred<"$_self.cast<mlir::quant::UniformQuantizedType>()" #
@@ -60,7 +60,7 @@ class UniformQuantizedSignedInt<int width>
   int bitwidth = width;
 }
 
-class UniformQuantizedUnsignedInt<int width>
+class StableHLO_UniformQuantizedUnsignedInt<int width>
   : Type<Or<[
       And<[CPred<"$_self.isa<mlir::quant::UniformQuantizedType>()">,
            CPred<"$_self.cast<mlir::quant::UniformQuantizedType>()" #
@@ -77,20 +77,20 @@ class UniformQuantizedUnsignedInt<int width>
   int bitwidth = width;
 }
 
-class UniformQuantizedSignedIntOfWidths<list<int> widths> :
-    AnyTypeOf<!foreach(w, widths, UniformQuantizedSignedInt<w>),
+class StableHLO_UniformQuantizedSignedIntOfWidths<list<int> widths> :
+    AnyTypeOf<!foreach(w, widths, StableHLO_UniformQuantizedSignedInt<w>),
               !interleave(widths, "/") # "-bit uniform quantized signed " #
               "integer">;
 
-class UniformQuantizedUnsignedIntOfWidths<list<int> widths> :
-    AnyTypeOf<!foreach(w, widths, UniformQuantizedUnsignedInt<w>),
+class StableHLO_UniformQuantizedUnsignedIntOfWidths<list<int> widths> :
+    AnyTypeOf<!foreach(w, widths, StableHLO_UniformQuantizedUnsignedInt<w>),
               !interleave(widths, "/") # "-bit uniform quantized unsigned " #
               "integer">;
 
 // Integer-based uniform quantized types. The definitions can be used to specify
 // operand's tensor types.
-def HLO_QuantizedSignedInt : UniformQuantizedSignedIntOfWidths<[4, 8, 16, 32]>;
-def HLO_QuantizedUnsignedInt : UniformQuantizedUnsignedIntOfWidths<[4, 8, 16, 32]>;
+def HLO_QuantizedSignedInt : StableHLO_UniformQuantizedSignedIntOfWidths<[4, 8, 16, 32]>;
+def HLO_QuantizedUnsignedInt : StableHLO_UniformQuantizedUnsignedIntOfWidths<[4, 8, 16, 32]>;
 def HLO_QuantizedInt :
     AnyTypeOf<[HLO_QuantizedSignedInt, HLO_QuantizedUnsignedInt]>;
